Changed datatype of id in Player.java and Persisten.java, added update, save, insert, delete in PlayerRepository

+++ Schnapsn/src/main/java/at/htlleonding/schnapsn/controller/Persistent.java
@@ -7,6 +7,6 @@ public interface Persistent<T> {
 
     public void insert(T entity);
 
-    public void delete(long id);
+    public void delete(int id);
 
 }
